% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-criterion.R
\name{add_criterion}
\alias{add_criterion}
\title{Add relative model fit criteria to model objects}
\usage{
add_criterion(
  x,
  criterion = c("loo", "waic"),
  overwrite = FALSE,
  save = TRUE,
  ...,
  r_eff = NA
)
}
\arguments{
\item{x}{A \link{measrfit} object.}

\item{criterion}{A vector of criteria to calculate and add to the model
object.}

\item{overwrite}{Logical. Indicates whether existing criteria should be
overwritten. For example, if LOO has already been added to the model object
\code{add_criterion()} is called again with \code{criterion = "loo"}, should LOO be
recalculated? Default is \code{FALSE}.}

\item{save}{Logical. Only relevant if a file was specified in the
\link{measrfit} object passed to \code{x}. If \code{TRUE} (the default), the model is
re-saved to the specified file when new criteria are added to the R object.
If \code{FALSE}, the new criteria will be added to the R object, but the saved
file will not be updated.}

\item{...}{Additional arguments passed to \code{\link[loo:loo]{loo::loo.array()}} or
\code{\link[loo:waic]{loo::waic.array()}}.}

\item{r_eff}{Vector of relative effective sample size estimates for the
likelihood (\code{exp(log_lik)}) of each observation. This is related to
the relative efficiency of estimating the normalizing term in
self-normalizing importance sampling when using posterior draws obtained
with MCMC. If MCMC draws are used and \code{r_eff} is not provided then
the reported PSIS effective sample sizes and Monte Carlo error estimates
will be over-optimistic. If the posterior draws are independent then
\code{r_eff=1} and can be omitted. The warning message thrown when \code{r_eff} is
not specified can be disabled by setting \code{r_eff} to \code{NA}. See the
\code{\link[loo:relative_eff]{relative_eff()}} helper functions for computing \code{r_eff}.}
}
\value{
A modified \link{measrfit} object with the \code{criteria} slot populated with
the specified criteria.
}
\description{
Add leave-one-out (LOO; Vehtari et al., 2017) and/or widely applicable
information criteria (WAIC; Watanabe, 2010) to fitted model objects.
}
\references{
Vehtari, A., Gelman, A., & Gabry, J. (2017). Practical Bayesian
model evaluation using leave-one-out cross-validation and WAIC.
\emph{Statistics and Computing, 27}(5), 1413-1432.
\url{https://doi.org/10.1007/s11222-016-9696-4}

Watanabe, S. (2010). Asymptotic equivalence of Bayes cross
validation and widely applicable information criterion in singular learning
theory. \emph{Journal of Machine Learning Research, 11}(116), 3571-3594.
\url{http://jmlr.org/papers/v11/watanabe10a.html}
}
