% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measr.R
\name{as.measr}
\alias{as.measr}
\title{Coerce to Measurement}
\usage{
as.measr(x)
}
\arguments{
\item{x}{The object to be coerced}
}
\value{
A \code{measr} object, or \code{NA} if it cannot be coerced.
}
\description{
\code{as.measr} will attempt to convert the specified object to a  
\code{measr} object.
}

