% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measrfit-class.R
\name{as_measrfit}
\alias{as_measrfit}
\alias{as_measrfit.default}
\title{Coerce objects to a \code{measrfit}}
\usage{
as_measrfit(x, class = character())

\method{as_measrfit}{default}(x, class = character())
}
\arguments{
\item{x}{An object to be coerced to a \code{measrfit}.}

\item{class}{Additional classes to be added (e.g., \code{measrdcm} for a
diagnostic classification model).}
}
\value{
An object of class \linkS4class{measrfit}.
}
\description{
Coerce objects to a \code{measrfit}
}
\examples{
\dontshow{if (measr_examples()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
rstn_mdm_lcdm <- measr_dcm(
  data = mdm_data, missing = NA, qmatrix = mdm_qmatrix,
  resp_id = "respondent", item_id = "item", type = "lcdm",
  method = "optim", seed = 63277, backend = "rstan"
)

new_obj <- as_measrfit(rstn_mdm_lcdm, class = "measrdcm")
\dontshow{\}) # examplesIf}
}
\seealso{
\linkS4class{measrfit}, \code{\link[=measrfit]{measrfit()}}
}
