% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{c.measrprior}
\alias{c.measrprior}
\title{Combine multiple measrprior objects into one measrprior}
\usage{
\method{c}{measrprior}(x, ..., replace = FALSE)
}
\arguments{
\item{x}{A \code{measrprior} object.}

\item{...}{Additional \code{measrprior} objects to be combined.}

\item{replace}{Should only unique priors be kept? If \code{TRUE}, the first prior
specified is kept.}
}
\value{
A \code{measrprior} object.
}
\description{
Combine multiple measrprior objects into one measrprior
}
