% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{default_dcm_priors}
\alias{default_dcm_priors}
\title{Default priors for the loglinear diagnostic classification model}
\usage{
default_dcm_priors(type = "lcdm")
}
\arguments{
\item{type}{Type of DCM to estimate. Must be one of
lcdm, dina, or dino.}
}
\value{
A \code{measrprior} object
}
\description{
Default priors for the loglinear diagnostic classification model
}
\examples{
default_dcm_priors(type = "lcdm")
}
