% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{default_lcdm_priors}
\alias{default_lcdm_priors}
\title{Default priors for the loglinear diagnostic classification model}
\usage{
default_lcdm_priors()
}
\value{
A \code{measrprior} object
}
\description{
Default priors for the loglinear diagnostic classification model
}
\examples{
default_lcdm_priors()
}
