% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difficulty.R
\name{difficulty}
\alias{difficulty}
\title{Item difficulty indices}
\usage{
difficulty(model)
}
\arguments{
\item{model}{The estimated model to be evaluated.}
}
\value{
A list with one element:
\itemize{
\item \code{weighted_pvalue}: A \link[tibble:tibble-package]{tibble} with one row
per item containing the weighted p-value for the item.
}
}
\description{
The item difficulty index is a measure of the difficulty of an item, often
represented as a proportion correct index or p-value (de la Torre & Douglas,
2008). Diagnostic classification models define p-values conditioned on latent
class, which provide class-level information (Thompson, 2019). The item-level
weighted p-value is a weighted average of class-level p-values, and defines
the likelihood that a randomly selected respondent correctly answers an item.
}
\details{
The item-level weighted p-value discrimination is calculated as the
average of class-level p-values (i.e. \eqn{\pi_{ic}}), weighted by the
relative frequency of each latent class (i.e. \eqn{\nu_c}).
}
\examples{
\dontshow{if (measr_examples()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
rstn_ecpe_lcdm <- measr_dcm(
  data = ecpe_data, missing = NA, qmatrix = ecpe_qmatrix,
  resp_id = "resp_id", item_id = "item_id", type = "lcdm",
  method = "optim", seed = 63277, backend = "rstan"
)

difficulty(rstn_ecpe_lcdm)
\dontshow{\}) # examplesIf}
}
\references{
de la Torre, J., & Douglas, J. A. (2008). Model evaluation and
multiple strategies in cognitive diagnosis: An analysis of fraction
subtraction data. \emph{Psychometrika, 73}, 595–624.
\doi{10.1007/s11336-008-9063-2}

Thompson, W. J. (2019). \emph{Bayesian psychometrics for diagnostic
assessments: A proof of concept} (Research Report No. 19-01). University
of Kansas; Accessible Teaching, Learning, and Assessment Systems.
\doi{10.35542/osf.io/jzqs8}
}
