% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measr.R
\name{is.measr}
\alias{is.measr}
\title{Coerc
Checks whether all the objects are of 'measr' type. Returns a logical vector.}
\usage{
is.measr(x)
}
\description{
Coerc
Checks whether all the objects are of 'measr' type. Returns a logical vector.
}

