% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{is.measrprior}
\alias{is.measrprior}
\title{Checks if argument is a \code{measrprior} object}
\usage{
is.measrprior(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical indicating if \code{x} is a \code{measrprior} object.
}
\description{
Checks if argument is a \code{measrprior} object
}
\examples{
prior1 <- prior(lognormal(0, 1), class = maineffect)
is.measrprior(prior1)

prior2 <- 3
is.measrprior(prior2)
}
