% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measrfit-class.R
\name{is_measrfit}
\alias{is_measrfit}
\title{Check if argument is a \code{measrfit} object}
\usage{
is_measrfit(x)
}
\arguments{
\item{x}{An object to be checked.}
}
\value{
A logical indicating is \code{x} is a \code{measrfit} object.
}
\description{
Check if argument is a \code{measrfit} object
}
\examples{
\dontshow{if (measr_examples()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
rstn_mdm_lcdm <- measr_dcm(
  data = mdm_data, missing = NA, qmatrix = mdm_qmatrix,
  resp_id = "respondent", item_id = "item", type = "lcdm",
  method = "optim", seed = 63277, backend = "rstan"
)

is_measrfit(rstn_mdm_lcdm)
\dontshow{\}) # examplesIf}
}
\seealso{
\linkS4class{measrfit}, \code{\link[=measrfit]{measrfit()}}, \code{\link[=as_measrfit]{as_measrfit()}}.
}
