% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo-methods.R
\name{loo_compare.measrfit}
\alias{loo_compare.measrfit}
\title{Relative model fit comparisons}
\usage{
\method{loo_compare}{measrfit}(x, ..., criterion = c("loo", "waic"), model_names = NULL)
}
\arguments{
\item{x}{A \link{measrfit} object.}

\item{...}{Additional objects of class \link{measrfit}.}

\item{criterion}{The name of the criterion to be extracted from the
\link{measrfit} object for comparison.}

\item{model_names}{}
}
\value{
The object returned by \code{\link[loo:loo_compare]{loo::loo_compare()}}.
}
\description{
A \code{\link[loo:loo_compare]{loo::loo_compare()}} method that is customized for \code{measrfit} objects. See
the \strong{loo} package \href{https://mc-stan.org/loo/articles/}{vignettes} for
details.
}
