% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measr.R
\name{measr}
\alias{\%pm\%}
\alias{measr}
\title{Create a \code{measr} Object}
\usage{
measr(val, err)

val \%pm\% err
}
\arguments{
\item{val}{The measured value}

\item{err}{The standard deviation of the measured value. This is usually
the measurement error}
}
\value{
A \code{measr} object with the given value and error. If one of the
parameters given cannot be coerced to a numeric object via
\code{\link{as.numeric}}, then \code{NA}s can be introduced in the return
value.
}
\description{
Creates a \code{measr} object with the given value and standard deviation.
}

