% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{measr_examples}
\alias{measr_examples}
\title{Determine if code is executed interactively or in pkgdown}
\usage{
measr_examples()
}
\value{
A logical value indicating whether or not the examples should be run.
}
\description{
Used for determining examples that shouldn't be run on CRAN, but can be run
for the pkgdown website.
}
\examples{
measr_examples()
}
