% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measrfit-class.R
\name{measrfit}
\alias{measrfit}
\title{Create a \code{measrfit} object}
\usage{
measrfit(
  data = list(),
  type = character(),
  prior = default_dcm_priors(type = type),
  stancode = character(),
  method = character(),
  algorithm = character(),
  backend = character(),
  model = NULL,
  respondent_estimates = list(),
  fit = list(),
  criteria = list(),
  reliability = list(),
  file = NULL,
  version = list(),
  class = character()
)
}
\arguments{
\item{class}{Additional classes to be added (e.g., \code{measrdcm} for a
diagnostic classification model).}
}
\value{
An \link{measrfit} object.
}
\description{
Models fitted with \strong{measr} are represented as a \code{measrfit} object. If a
model is estimated with \emph{Stan}, but not \strong{measr}, a \code{measrfit} object can be
created in order to access other functionality in \strong{measr} (e.g., model fit,
reliability).
}
\seealso{
\link{measrfit} \code{\link[=measr_dcm]{measr_dcm()}}
}
