% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R, R/zzz-methods-loo-waic.R,
%   R/zzz-methods-m2.R, R/zzz-methods-posterior.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{dcm_specify}
\alias{create_profiles}
\alias{get_parameters}
\alias{prior}
\alias{default_dcm_priors}
\alias{lcdm}
\alias{dina}
\alias{dino}
\alias{crum}
\alias{unconstrained}
\alias{independent}
\alias{hdcm}
\alias{E}
\alias{Pr}
\alias{rvar_median}
\alias{rvar_sum}
\alias{rvar_prod}
\alias{rvar_min}
\alias{rvar_max}
\alias{rvar_mean}
\alias{rvar_sd}
\alias{rvar_var}
\alias{rvar_mad}
\alias{loo}
\alias{waic}
\alias{loo_compare}
\alias{fit_m2}
\alias{as_draws}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dcm2}{\code{\link[dcm2]{fit_m2}}}

  \item{dcmstan}{\code{\link[dcmstan]{create_profiles}}, \code{\link[dcmstan:measurement-model]{crum}}, \code{\link[dcmstan]{dcm_specify}}, \code{\link[dcmstan]{default_dcm_priors}}, \code{\link[dcmstan:measurement-model]{dina}}, \code{\link[dcmstan:measurement-model]{dino}}, \code{\link[dcmstan]{get_parameters}}, \code{\link[dcmstan:structural-model]{hdcm}}, \code{\link[dcmstan:structural-model]{independent}}, \code{\link[dcmstan:measurement-model]{lcdm}}, \code{\link[dcmstan]{prior}}, \code{\link[dcmstan:structural-model]{unconstrained}}}

  \item{loo}{\code{\link[loo]{loo}}, \code{\link[loo]{loo_compare}}, \code{\link[loo]{waic}}}

  \item{posterior}{\code{\link[posterior:draws]{as_draws}}, \code{\link[posterior:rvar-summaries-over-draws]{E}}, \code{\link[posterior:rvar-summaries-over-draws]{Pr}}, \code{\link[posterior:rvar-summaries-within-draws]{rvar_mad}}, \code{\link[posterior:rvar-summaries-within-draws]{rvar_max}}, \code{\link[posterior:rvar-summaries-within-draws]{rvar_mean}}, \code{\link[posterior:rvar-summaries-within-draws]{rvar_median}}, \code{\link[posterior:rvar-summaries-within-draws]{rvar_min}}, \code{\link[posterior:rvar-summaries-within-draws]{rvar_prod}}, \code{\link[posterior:rvar-summaries-within-draws]{rvar_sd}}, \code{\link[posterior:rvar-summaries-within-draws]{rvar_sum}}, \code{\link[posterior:rvar-summaries-within-draws]{rvar_var}}}
}}

