% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-methods-yens-q.R
\name{yens_q}
\alias{yens_q}
\title{Yen's Q statistic for diagnostic models}
\usage{
yens_q(x, ..., crit_value = 0.2, force = FALSE)
}
\arguments{
\item{x}{A \link[=dcm_estimate]{measrdcm} object.}

\item{...}{Unused.}

\item{crit_value}{Numerical. A critical value threshold for evaluating the
residual correlations from the Yen's Q calculations.}

\item{force}{If the Yen's Q information has already been added to the model
object with \code{\link[=yens_q]{yens_q()}}, should it be recalculated? Default is
\code{FALSE}.}
}
\value{
A tibble with the residual correlation and flags for all item pairs.
}
\description{
Calculate the Yen's Q statistic for diagnostic models.
}
\examples{
\dontshow{if (measr_examples()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# example code
model_spec <- dcm_specify(qmatrix = dcmdata::mdm_qmatrix,
                          identifier = "item")
model <- dcm_estimate(dcm_spec = model_spec, data = dcmdata::mdm_data,
                      identifier = "respondent", method = "optim",
                      seed = 63277)

yens_q(model)
\dontshow{\}) # examplesIf}
}
\references{
Chen, W.-H., & Thissen, D. (1997). Local dependence indexes for
item pairs using item response theory. \emph{Journal of Educational and
Behavioral Statistics, 22}(3), 265-389. \doi{10.3102/10769986022003265}

Christensen, K. B., Makransky, G., & Horton, M. (2016). Critical
values for Yen's Q3: Identification of local dependence in the Rasch model
using residual correlations. \emph{Applied Psychological Measurement, 41}(3),
178-194. \doi{10.1177/0146621616677520}
}
