% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-methods-bayes-factor.R
\name{bayes_factor}
\alias{bayes_factor}
\alias{add_marginal_likelihood}
\title{Bayes factor for diagnostic models}
\usage{
bayes_factor(x, ..., y, force = FALSE)

add_marginal_likelihood(x)
}
\arguments{
\item{x}{A \link[=dcm_estimate]{measrdcm} object.}

\item{...}{Unused.}

\item{y}{A \link[=dcm_estimate]{measrdcm} object.}

\item{force}{If the criterion has already been added to the
model object with \code{\link[=add_criterion]{add_criterion()}}, should it be recalculated. Default is
\code{FALSE}.}
}
\value{
A modified \link[=dcm_estimate]{measrdcm} object with the Bayes factor
slot populated with the Bayes factor statistic.
}
\description{
Calculate Bayes factors for diagnostic models. The results of the Bayes
factor calculations are designed to be consistent with the reporting
guidelines for Bayesian analysis developedy by Kruschke (2021).
}
\examples{
\dontshow{if (measr_examples()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# example code
dina_spec <- dcmstan::dcm_specify(qmatrix = q_matrix[, -1],
                                  measurement_model = dina())
dino_spec <- dcmstan::dcm_specify(qmatrix = q_matrix, identifier = "item",
                                  measurement_model = dino())
rstn_dina <- dcm_estimate(dina_spec, data = dina_data,
                          identifier = "resp_id", method = "optim",
                          backend = "rstan", seed = 63277,
                          precompiled = stanmodels$test_dina)
rstn_dino <- dcm_estimate(dino_spec, data = dino_data,
                          identifier = "resp_id", method = "optim",
                          backend = "rstan", seed = 63277,
                          precompiled = stanmodels$test_dina)
rstn_dina <- add_respondent_estimates(rstn_dina)
rstn_dino <- add_respondent_estimates(rstn_dino)

bayes_factor(rstn_dina, rstn_dino)
\dontshow{\}) # examplesIf}
}
\references{
Kruschke, J. K. (2021). Bayesian analysis reporting guidelines.
\emph{Nature, 5}, 1282-1291. \doi{10.1038/s41562-021-01177-7}
}
