% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measr.R
\name{is.measr}
\alias{is.measr}
\title{Is an Object a Measured Value?}
\usage{
is.measr(x)
}
\arguments{
\item{x}{The object to be tested}
}
\value{
\code{TRUE} if the object is a \code{measr} object. Otherwise, 
    returns \code{FALSE}.
}
\description{
Checks whether the object is of \code{measr} type.
}

