% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-loo.R
\name{loglik_array}
\alias{loglik_array}
\alias{loglik_array.measrdcm}
\title{Extract the log-likelihood of an estimated model}
\usage{
loglik_array(model)

\method{loglik_array}{measrdcm}(model)
}
\arguments{
\item{model}{A \link{measrfit} object.}
}
\value{
A "\code{\link[posterior:draws_array]{draws_array}}" object containing the
log-likelihood estimates for the model.
}
\description{
The \code{loglik_array()} methods for \link{measrfit} objects calculates the
log-likelihood for an estimated model via the generated quantities
functionality in \emph{Stan} and returns the draws of the \code{log_lik} parameter.
}
