% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measr-package.R
\docType{package}
\name{measr-package}
\alias{measr}
\alias{measr-package}
\title{measr: Bayesian Psychometric Measurement Using 'Stan'}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wjakethompson/measr}
  \item Report bugs at \url{https://github.com/wjakethompson/measr/issues}
}

}
\author{
\strong{Maintainer}: W. Jake Thompson \email{wjakethompson@gmail.com} (\href{https://orcid.org/0000-0001-7339-0300}{ORCID})

Other contributors:
\itemize{
  \item University of Kansas Center for Research [copyright holder]
  \item Institute of Education Sciences [funder]
}

}
\keyword{internal}
