% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measr-package.R
\docType{package}
\name{measr-package}
\alias{measr}
\alias{measr-package}
\title{measr: Bayesian Psychometric Measurement Using 'Stan'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Estimate diagnostic classification models (also called cognitive diagnostic models) with 'Stan'. Diagnostic classification models are confirmatory latent class models, as described by Rupp et al. (2010, ISBN: 978-1-60623-527-0). Automatically generate 'Stan' code for the general loglinear cognitive diagnostic diagnostic model proposed by Henson et al. (2009) \doi{10.1007/s11336-008-9089-5} and other subtypes that introduce additional model constraints. Using the generated 'Stan' code, estimate the model evaluate the model's performance using model fit indices, information criteria, and reliability metrics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://measr.info}
  \item \url{https://github.com/wjakethompson/measr}
  \item Report bugs at \url{https://github.com/wjakethompson/measr/issues}
}

}
\author{
\strong{Maintainer}: W. Jake Thompson \email{wjakethompson@gmail.com} (\href{https://orcid.org/0000-0001-7339-0300}{ORCID})

Other contributors:
\itemize{
  \item Nathan Jones \email{jonesnateb@gmail.com} (\href{https://orcid.org/0000-0001-6177-7161}{ORCID}) [contributor]
  \item Matthew Johnson (Provided code adapted for reliability.measrdcm()) [copyright holder]
  \item Paul-Christian Bürkner (Author of eval_silent()) [copyright holder]
  \item University of Kansas [copyright holder]
  \item Institute of Education Sciences [funder]
}

}
\keyword{internal}
