% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measr.R
\name{measr}
\alias{\%pm\%}
\alias{measr}
\title{Create a \code{measr} object}
\usage{
measr(val, err)

val \%pm\% err
}
\arguments{
\item{val}{The measured value}

\item{err}{The standard deviation of the measured value. This is usually
the measurement error}
}
\value{
A \code{measr} object with the given value and error.
}
\description{
Creates a \code{measr} object with the given value and standard deviation.
}

