% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-dcm.R
\name{measr_dcm}
\alias{measr_dcm}
\title{Fit Bayesian diagnostic classification models}
\usage{
measr_dcm(
  data,
  resp_id = NULL,
  missing = NA,
  qmatrix,
  type = c("lcdm"),
  method = c("mcmc", "optim"),
  prior = NULL,
  backend = getOption("measr.backend", "rstan"),
  ...
)
}
\arguments{
\item{data}{Response data. A data frame with 1 row per respondent and 1
column per item.}

\item{resp_id}{Optional. Variable name of a column in \code{data} that
contains respondent identifiers. \code{NULL} (the default) indicates that no
identifiers are present in the data, and row numbers will be used as
identifiers.}

\item{missing}{An R expression specifying how missing data is coded (e.g.,
\code{NA}, \code{"."}, \code{-99}, etc.). The default is \code{NA}.}

\item{qmatrix}{The Q-matrix. A data frame with 1 row per item and 1 column
per attribute. All cells should be either 0 (item does not measure the
attribute) or 1 (item does measure the attribute).}

\item{type}{Type of DCM to estimate. Must be one of
lcdm, dina, or dino)`.}

\item{method}{Estimation method. Options are \code{"mcmc"}, which uses Stan's
sampling method, or \code{"optim"}, which uses Stan's optimizer.}

\item{prior}{A \link[=measrprior]{measrprior} object. If \code{NULL}, default priors
are used, as specified by \code{\link[=default_dcm_priors]{default_dcm_priors()}}.}

\item{backend}{Character string naming the package to use as the backend for
fitting the Stan model. Options are "rstan" (the default) or "cmdstanr".
Can be set globally for the current R session via the "measr.backend"
option (see options). Details on the rstan and cmdstanr packages are
available at \url{https://mc-stan.org/rstan/} and
\url{https://mc-stan.org/cmdstanr/}, respectively.}

\item{...}{Additional arguments passed to Stan.
\itemize{
\item For \code{backend = "rstan"}, arguments are passed to \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}}
or \code{\link[rstan:stanmodel-method-optimizing]{rstan::optimizing()}}.
\item For \code{backend = "cmdstanr"}, arguments are passed to the
\code{\link[cmdstanr:model-method-sample]{$sample()}} or
\code{\link[cmdstanr:model-method-optimize]{$optimize()}} methods of the
\link[cmdstanr]{CmdStanModel} class.
}}
}
\value{
A measrfit object.
}
\description{
Something...
}
