% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{measr_extract}
\alias{measr_extract}
\alias{measr_extract.measrdcm}
\title{Extract components of a \code{measrfit} object.}
\usage{
measr_extract(model, ...)

\method{measr_extract}{measrdcm}(model, what, ...)
}
\arguments{
\item{model}{The estimated to extract information from.}

\item{...}{Additional arguments passed to specific methods.}

\item{what}{Character string. The information to be extracted. See details
for available options.}
}
\value{
The extracted information. The specific structure will vary depending
on what is being extracted, but usually the returned object is a
\link[tibble:tibble-package]{tibble} with the requested information.
}
\description{
Extract components of a \code{measrfit} object.

Extract components of an estimated diagnostic classification model
}
\details{
For diagnostic classification models, we can extract the following
information:
\itemize{
\item \code{item_param}: The estimated item parameters. This shows the name of the
parameter, the class of the parameter, and the estimated value.
\item \code{strc_param}: The estimated structural parameters. This is base rate of
membership in each class. This shows the class pattern and the estimated
proportion of respondents in each class.
\item \code{prior}: The priors used when estimating the model.
}
}
\section{Methods (by class)}{
\itemize{
\item \code{measr_extract(measrdcm)}: Extract components of an estimated diagnostic
classification model.

}}
\examples{
\dontshow{if (measr_examples()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
rstn_mdm_lcdm <- measr_dcm(
  data = mdm_data, missing = NA, qmatrix = mdm_qmatrix,
  resp_id = "respondent", item_id = "item", type = "lcdm",
  method = "optim", seed = 63277, backend = "rstan"
)

extract(rstn_mdm_lcdm, "strc_param")
\dontshow{\}) # examplesIf}
}
