% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-methods-qmatrix-validation.R
\name{qmatrix_validation}
\alias{qmatrix_validation}
\title{Add Q-matrix validation model objects}
\usage{
qmatrix_validation(x, ..., epsilon = 0.95, force = FALSE)
}
\arguments{
\item{x}{A \link[=dcm_estimate]{measrdcm} object.}

\item{...}{Unused.}

\item{epsilon}{The threshold for proportion of variance accounted for to flag
items for appropriate empirical specifications. The default is .95 as
implemented by de la Torre and Chiu (2016).}

\item{force}{If the criterion has already been added to the
model object with \code{\link[=add_criterion]{add_criterion()}}, should it be recalculated. Default is
\code{FALSE}.}
}
\value{
A tibble containing the estimated item-level discrimination indices
used to empirically validate the Q-matrix.
}
\description{
Add Q-matrix validation metrics to fitted model objects. This function is a
wrapper around other functions to compute the metrics. The benefit of
using this wrapper is that the Q-matrix validation metrics are saved as part
of the model object so that time-intensive calculations do not need to be
repeated.
}
\examples{
\dontshow{if (measr_examples()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
mod_spec <- dcm_specify(qmatrix = dcmdata::ecpe_qmatrix,
                        measurement_model = dcmstan::lcdm(),
                        structural_model =
                          dcmstan::hdcm(hierarchy =
                                          paste0("lexical -> cohesive -> ",
                                                 "morphosyntactic")),
                        identifier = "item_id")
rstn_ecpe <- dcm_estimate(mod_spec, data = dcmdata::ecpe_data,
                          identifier = "resp_id", backend = "rstan",
                          method = "optim")
rstn_ecpe <- add_respondent_estimates(rstn_ecpe)
q_matrix_validation <- qmatrix_validation(rstn_ecpe)
\dontshow{\}) # examplesIf}
}
\references{
de la Torre, J., & Chiu, C.-Y. (2016). A general method of
empirical Q-matrix validation. \emph{Psychometrika, 81}(2), 253-273.
\doi{10.1007/s11336-015-9467-8}
}
