% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmatrix-validation.R
\name{qmatrix_validation}
\alias{qmatrix_validation}
\alias{add_qmatrix_validation}
\title{Add Q-matrix validation model objects}
\usage{
add_qmatrix_validation(mod, epsilon = 0.95, overwrite = FALSE, save = TRUE)
}
\arguments{
\item{mod}{A \link{measrfit} object.}

\item{epsilon}{The threshold for proportion of variance accounted for to flag
items for appropriate empirical specifications. The default is .95 as
implemented by de la Torre and Chiu (2016).}

\item{overwrite}{Logical. Indicates whether specified elements that have
already been added to the estimated model should be overwritten. Default is
\code{FALSE}.}

\item{save}{Logical. Only relevant if a file was specified in the
\link{measrfit} object passed to \code{x}. If \code{TRUE} (the default), the model is
re-saved to the specified file when new criteria are added to the \code{R}
object. If \code{FALSE}, the new criteria will be added to the \code{R} object, but
the saved file will not be updated.}
}
\value{
A modified \link{measrfit} object with the estimated item-level
discrimination index are added to the \verb{$qmatrix_validation} element of the
fitted model.
}
\description{
Add Q-matrix validation metrics to fitted model objects. This function is a
wrapper around other functions to compute the metrics. The benefit of
using this wrapper is that the Q-matrix validation metrics are saved as part
of the model object so that time-intensive calculations do not need to be
repeated.
}
\examples{
\dontshow{if (measr_examples()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
cmds_ecpe_lcdm <- measr_dcm(
  data = ecpe_data, missing = NA, qmatrix = ecpe_qmatrix,
  resp_id = "resp_id", item_id = "item_id", type = "lcdm",
  method = "optim", seed = 63277, backend = "cmdstanr",
  prior = c(prior(uniform(-15, 15), class = "intercept"),
            prior(uniform(0, 15), class = "maineffect"),
            prior(uniform(-15, 15), class = "interaction"))
)

cmds_ecpe_lcdm <- add_qmatrix_validation(cmds_ecpe_lcdm)
\dontshow{\}) # examplesIf}
}
\references{
de la Torre, J., & Chiu, C.-Y. (2016). A general method of
empirical Q-matrix validation. \emph{Psychometrika, 81}(2), 253-273.
\doi{10.1007/s11336-015-9467-8}
}
