% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo-methods.R, R/m2-methods.R,
%   R/posterior-methods.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{loo}
\alias{waic}
\alias{loo_compare}
\alias{fit_m2}
\alias{as_draws}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dcm2}{\code{\link[dcm2]{fit_m2}}}

  \item{loo}{\code{\link[loo]{loo}}, \code{\link[loo]{loo_compare}}, \code{\link[loo]{waic}}}

  \item{posterior}{\code{\link[posterior:draws]{as_draws}}}
}}

