% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo-methods.R, R/m2-methods.R,
%   R/posterior-methods.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{loo}
\alias{waic}
\alias{loo_compare}
\alias{fit_m2}
\alias{as_draws}
\alias{E}
\alias{Pr}
\alias{rvar_median}
\alias{rvar_sum}
\alias{rvar_prod}
\alias{rvar_min}
\alias{rvar_max}
\alias{rvar_mean}
\alias{rvar_sd}
\alias{rvar_var}
\alias{rvar_mad}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dcm2}{\code{\link[dcm2]{fit_m2}}}

  \item{loo}{\code{\link[loo]{loo}}, \code{\link[loo]{loo_compare}}, \code{\link[loo]{waic}}}

  \item{posterior}{\code{\link[posterior:draws]{as_draws}}, \code{\link[posterior:rvar-summaries-over-draws]{E}}, \code{\link[posterior:rvar-summaries-over-draws]{Pr}}, \code{\link[posterior:rvar-summaries-within-draws]{rvar_mad}}, \code{\link[posterior:rvar-summaries-within-draws]{rvar_max}}, \code{\link[posterior:rvar-summaries-within-draws]{rvar_mean}}, \code{\link[posterior:rvar-summaries-within-draws]{rvar_median}}, \code{\link[posterior:rvar-summaries-within-draws]{rvar_min}}, \code{\link[posterior:rvar-summaries-within-draws]{rvar_prod}}, \code{\link[posterior:rvar-summaries-within-draws]{rvar_sd}}, \code{\link[posterior:rvar-summaries-within-draws]{rvar_sum}}, \code{\link[posterior:rvar-summaries-within-draws]{rvar_var}}}
}}

