% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo-methods.R
\name{waic.measrfit}
\alias{waic.measrfit}
\title{Widely applicable information criterion (WAIC)}
\usage{
\method{waic}{measrfit}(x, ...)
}
\arguments{
\item{x}{A \link{measrfit} object.}

\item{...}{Additional arguments passed to \code{\link[loo:waic]{loo::waic.array()}}.}
}
\value{
The object returned by \code{\link[loo:waic]{loo::waic.array()}}.
}
\description{
A \code{\link[loo:waic]{loo::waic()}} method that is customized for \code{measrfit} objects. This is a
simple wrapper around \code{\link[loo:waic]{loo::waic.array()}}. See the \strong{loo} package
\href{https://mc-stan.org/loo/articles/}{vignettes} for details.
}
